[prometheus-blackbox-exporter] Add custom relabeling and quote values (#2033)
* [blackbox] Add custom relabeling and quote values

Adds `additionalRelabeling` to ServiceMonitor so that custom relabeling
can be done. I wanted this because I wanted to be able to add the module
as a label but I think there are other good reasons to have it.

While I was here I added `quote` for `additionalMetricsRelabels` so that
values wouldn't be converted.

Signed-off-by: Evan Pitstick <epitstick@paloaltonetworks.com>

* bumping minor version

Signed-off-by: Evan Pitstick <epitstick@paloaltonetworks.com>
Signed-off-by: Aidan Jensen <aidan@artificial.com>